created short term scheduler

+++ Kernel/include/kernel_configuration.h
@@ -10,7 +10,7 @@ char* get_cpu_dispatch_port();
 char* get_cpu_interrupt_port();
 char* get_listening_port();
 char* get_scheduling_algorithm();
-uint32_t get_initial_estimation();
+double get_initial_estimation();
 double get_alpha();
 uint32_t get_multiprogramming_degree();
 uint32_t get_max_block_time();
